mainV2.c
allows user to enter an input file, it takes that file and reads it line by line  storing anything in front of a comma in an array. Then it sorts out the data, specifically takes the original array and "deletes" entries that have negative drag area and stores the positive entries in a new array.